Rearrange some types to be better separated
