+++ src/react/Computed.tsx
@@ -1,6 +1,6 @@
 import { computeSelector } from '@legendapp/state';
 import { ReactElement, ReactNode } from 'react';
-import type { ObservableParam } from '../observableTypes';
+import type { ObservableParam } from '@legendapp/state';
 import { useSelector } from './useSelector';
 
 export function Computed({ children }: { children: ObservableParam | (() => ReactNode) }): ReactElement {
