+++ src/react/reactInterfaces.ts
@@ -1,6 +1,5 @@
+import type { Observable, Selector } from '@legendapp/state';
 import type { FC, LegacyRef, ReactNode } from 'react';
-import type { Observable } from '../observableTypes';
-import type { Selector } from '../observableInterfaces';
 
 export type ShapeWithNew$<T> = Partial<Omit<T, 'children'>> & {
     [K in keyof T as K extends `$${string & K}` ? K : `$${string & K}`]?: Selector<T[K]>;
