+++ src/retry.ts
@@ -1,6 +1,5 @@
 import { whenReady } from './when';
-import type { NodeValue } from './observableInterfaces';
-import type { RetryOptions } from './sync/syncTypes';
+import type { NodeValue, RetryOptions } from './observableInterfaces';
 import { isPromise } from './is';
 
 function calculateRetryDelay(retryOptions: RetryOptions, attemptNum: number): number | null {
